Fix test_ArgDefaults_function_with_required_named test on Windows
The changes in addSource are so that the driver gets the convertedPath
(newFile already converts).

The change in the test is to format a path for importing on Windows - it
seems that importing a file with an absolute path is a bit wonky on
Windows (I presume since this is testing imports it's using absolute
paths for a reason).

The helper in the Mixin is for doing the conversion from a Windows path
to an importable version of it (this has many questions to answer and
maybe already exists somewhere?).

See #32095.

Change-Id: I9732bab89ab17fc2eed51b14ba4c4651d5a48f26